remove tree-based repository, service, and entity classes; refactor manuscript info part for branch token count type conversion

+++ marginalia/src/main/java/com/github/enerccio/marginalia/ui/dialogs/manuscript/ManuscriptInfoPart.java
@@ -184,12 +184,12 @@ public class ManuscriptInfoPart implements ManuscriptDialogPart {
         int totalWords = chatMessageService.getTotalWordCount(manuscript);
         int totalTokens = chatMessageService.getTotalTokenCount(manuscript);
         int branchWords = chatMessageService.getBranchWordCount(manuscript.getActiveLeaf());
-        int branchTokens = chatMessageService.getBranchTokenCount(manuscript.getActiveLeaf());
+        long branchTokens = chatMessageService.getBranchTokenCount(manuscript.getActiveLeaf());
 
         totalWordCountField.setValue(totalWords);
         totalTokenCountField.setValue(totalTokens);
         branchWordCountField.setValue(branchWords);
-        branchTokenCountField.setValue(branchTokens);
+        branchTokenCountField.setValue((int) branchTokens);
     }
 
     @Override
